requirements.txt
pandas
duckdb
PyMuPDF
sentence-transformers==3.0.0 # Updated to the latest stable version
qdrant-client==1.7.0
sqlalchemy==2.0.23
streamlit
peft
transformers==4.41.0 # Updated to a recent version compatible with sentence-transformers 3.0.0
python-dotenv
langchain-google-genai==0.0.11
duckdb-engine
torch==2.2.0
huggingface-hub==0.23.0 # Updated to a recent version compatible with transformers 4.41.0
langchain==0.1.16
langchain-core==0.1.51


